refactor(fix some const, float, and refs, and style changes)

+++ libs/shapes/DeformedLayeredCircle.h
@@ -7,7 +7,7 @@ namespace soo {
 class DeformedLayeredCircle {
  public:
   DeformedLayeredCircle() = default;
-  DeformedLayeredCircle(const ofVec2f& center, const unsigned long num_layers,
+  DeformedLayeredCircle(const ofVec2f& center, const int num_layers,
                         const float spacing_between_layers,
                         std::vector<float>& radii);
